app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
pydantic-settings
httpx
pytest
